TGMC Final Commit - Fixes Ammo Count + symbols + Medigun/Curabitur
Changes:
- Code copied over to hud.dm and commented out, to prep for upstream pull.
- Commented-out, unused code removed.
- CHOMPEdits removed, pending upstream PR.
- Medigun no longer shows 2x the actual ammo loaded.
- Curabitur now fetches the amount of shots correctly, bar 1 extra for some reason.
- Curabitur ammo updates on automatic recharge ticks.
- A ton of projectile get_ammo_count optimization - mostly swaps the massive if/else for for loops. Fixes multiple guns not getting ammo properly, now casings won't register as ammo unless they have a shell in the casing.
- hud_state_empty defined as base on all bullets.
- Corrects energy weapons missing a hud_state_empty.
- Gives Ion weapons a hud_state_empty

All changes have been tested, and bar the Curabitur having an extra shot for ??? reason, which is a base bug I've yet to fix, this is entirely ready for testmerge.

+++ code/_onclick/hud/hud_ch.dm
@@ -1,5 +1,6 @@
 /**
  * CHOMP-Specific HUD override. This will be backported to Polaris if they want it. 
+ * Comment this out once Polaris has it!
  */
 ///Add an ammo hud to the user informing of the ammo count of G
 /datum/hud/proc/add_ammo_hud(mob/living/user, obj/item/weapon/gun/G)
@@ -14,6 +15,8 @@
 ///Remove the ammo hud related to the gun G from the user
 /datum/hud/proc/remove_ammo_hud(mob/living/user, obj/item/weapon/gun/G)
 	var/obj/screen/ammo/ammo_hud = ammo_hud_list[G]
+	if(isnull(ammo_hud))
+		return
 	ammo_hud.remove_hud(user, G)
 	qdel(ammo_hud)
 	ammo_hud_list -= G
